Summarizing an evpost object

summary method for class "evpost"

# S3 method for evpost
summary(object, add_pu = FALSE, ...)

Arguments

object
An object of class "evpost", a result of a call to rpost or rpost_rcpp.

add_pu
Includes in the summary of the simulated values the threshold exceedance probability \(p\). Only relevant when model == "bingp" was used in the call to rpost or rpost_rcpp.

...
Additional arguments passed on to print.

Value

Prints

information about the ratio-of-uniforms bounding box, i.e. object$box

an estimate of the probability of acceptance, i.e. object$pa

a summary of the simulated values, via summary(object$sim_vals)

See also

Examples

# GP posterior
data(gom)
u <- stats::quantile(gom, probs = 0.65)
fp <- set_prior(prior = "flat", model = "gp", min_xi = -1)
gpg <- rpost_rcpp(n = 1000, model = "gp", prior = fp, thresh = u,
                  data = gom)
summary(gpg)

#> ru bounding box:  
#>                box       vals1        vals2 conv
#> a        1.0000000  0.00000000  0.000000000    0
#> b1minus -0.1477092 -0.23028294 -0.005717161    0
#> b2minus -0.1397605  0.07992047 -0.212484993    0
#> b1plus   0.1913469  0.34032615 -0.015695239    0
#> b2plus   0.2034900  0.10510888  0.373501793    0
#> 
#> estimated probability of acceptance:  
#> [1] 0.5288207
#> 
#> sample summary 
#>     sigma[u]           xi          
#>  Min.   :1.062   Min.   :-0.12185  
#>  1st Qu.:1.576   1st Qu.: 0.06267  
#>  Median :1.729   Median : 0.13276  
#>  Mean   :1.740   Mean   : 0.14031  
#>  3rd Qu.:1.892   3rd Qu.: 0.20396  
#>  Max.   :2.620   Max.   : 0.60221