Simulation from a Dirichlet distribution

Simulates from a Dirichlet distribution with concentration parameter vector \(\alpha\) = (\(\alpha_1\), ..., \(\alpha_K\)).

rDir(n = 1, alpha = c(1, 1))

Arguments

n
A numeric scalar. The size of sample required.

alpha
A numeric vector. Dirichlet concentration parameter.

Value

An n by length(alpha) numeric matrix.

Details

The simulation is based on the property that if \(Y_1, \ldots, Y_K\) are independent, \(Y_i\) has a gamma(\(\alpha_i\), 1) distribution and \(S = Y_1 + \cdots + Y_k\) then \((Y_1, \ldots, Y_K) / S\) has a Dirichlet(\(\alpha_1\), ..., \(\alpha_K\)) distribution.

See https://en.wikipedia.org/wiki/Dirichlet_distribution#Gamma_distribution

References

Kotz, S., Balakrishnan, N. and Johnson, N. L. (2000) Continuous Multivariate Distributions, vol. 1, Models and Applications, 2nd edn, ch. 49. New York: Wiley. https://doi.org/10.1002/0471722065

See also

Examples

rDir(n = 10, alpha = 1:4)

#>              [,1]       [,2]      [,3]      [,4]
#>  [1,] 0.207146679 0.14854489 0.1320263 0.5122822
#>  [2,] 0.009740641 0.17294038 0.4681217 0.3491973
#>  [3,] 0.218151392 0.14324434 0.3827741 0.2558302
#>  [4,] 0.023516805 0.07520053 0.4140892 0.4871935
#>  [5,] 0.023084137 0.49731355 0.2108664 0.2687360
#>  [6,] 0.008039905 0.24069465 0.3213728 0.4298927
#>  [7,] 0.036956932 0.16191880 0.3474666 0.4536577
#>  [8,] 0.016962455 0.20462139 0.4091748 0.3692414
#>  [9,] 0.203186377 0.18946611 0.1529401 0.4544075
#> [10,] 0.041056408 0.37833740 0.2859789 0.2946273