Summarizing an evpost object

summary method for class "evpost"

# S3 method for evpost
summary(object, add_pu = FALSE, ...)

Arguments

object

An object of class "evpost", a result of a call to rpost or rpost_rcpp.

add_pu

Includes in the summary of the simulated values the threshold exceedance probability \(p\). Only relevant when model == "bingp" was used in the call to rpost or rpost_rcpp.

...

Additional arguments passed on to print.

Value

Prints

information about the ratio-of-uniforms bounding box, i.e. object$box

an estimate of the probability of acceptance, i.e. object$pa

a summary of the simulated values, via summary(object$sim_vals)

See also

Examples

# GP posterior
u <- stats::quantile(gom, probs = 0.65)
fp <- set_prior(prior = "flat", model = "gp", min_xi = -1)
gpg <- rpost_rcpp(n = 1000, model = "gp", prior = fp, thresh = u,
                  data = gom)
summary(gpg)
#> ru bounding box:  
#>                box       vals1       vals2 conv
#> a        1.0000000  0.00000000  0.00000000    0
#> b1minus -0.1477092 -0.23028294 -0.00571716    0
#> b2minus -0.1397605  0.07992048 -0.21248500    0
#> b1plus   0.1913469  0.34032616 -0.01569524    0
#> b2plus   0.2034900  0.10510904  0.37350176    0
#> 
#> estimated probability of acceptance:  
#> [1] 0.5015045
#> 
#> sample summary 
#>     sigma[u]           xi          
#>  Min.   :1.006   Min.   :-0.12354  
#>  1st Qu.:1.569   1st Qu.: 0.05992  
#>  Median :1.736   Median : 0.12749  
#>  Mean   :1.742   Mean   : 0.13704  
#>  3rd Qu.:1.902   3rd Qu.: 0.19885  
#>  Max.   :2.634   Max.   : 0.56799